Introduced a map method to make the null check bug more clear

+++ null-check-on-source/src/main/java/com/bol/nullcheck/ATarget.java
@@ -8,13 +8,13 @@ public class ATarget {
     public String getS() {
         return s;
     }
-    private final Integer i;
+    private final Double i;
 
-    public Integer getI() {
+    public Double getI() {
         return i;
     }
 
-    public ATarget(String s, Integer i) {
+    public ATarget(String s, Double i) {
         this.s = s;
         this.i = i;
     }
